refactor: move interface definition to one file
