--- src/router/interface.ts
@@ -1,9 +1,0 @@
-export interface QueryInterface {
-  base: string
-  quote: string
-  side: 'BUY' | 'SELL'
-  amount?: number
-  feefactor?: number
-  uniqId?: number | string
-  userAddr?: string
-}
